requirements.txt
streamlit==1.32.0
pandas==2.2.0
numpy==1.26.4
plotly==5.18.0
scikit-learn==1.4.0
python-dotenv==1.0.1
openai==1.12.0
requests==2.31.0
folium==0.15.1
streamlit-folium==0.15.0 